Moved select_time to calendar and refactored for less redundancy
