+++ xclim/core/utils.py
@@ -6,7 +6,6 @@ Miscellaneous indices utilities
 
 Helper functions for the indices computation, indicator construction and other things.
 """
-import datetime as pydt
 import logging
 import os
 import warnings
@@ -18,26 +17,16 @@ from importlib.resources import open_text
 from inspect import Parameter
 from pathlib import Path
 from types import FunctionType
-from typing import Any, Callable, NewType, Optional, Sequence, Tuple, Union
+from typing import Callable, NewType, Optional, Sequence, Tuple, Union
 
-import cftime
 import numpy as np
-import pandas as pd
 import xarray as xr
 from boltons.funcutils import update_wrapper
 from dask import array as dsk
-from xarray import CFTimeIndex, DataArray, Dataset
-from xarray.coding.cftime_offsets import to_cftime_datetime
+from xarray import DataArray, Dataset
 from yaml import safe_dump, safe_load
 
 logger = logging.getLogger("xclim")
-
-
-# Names of calendars that have the same number of days for all years
-uniform_calendars = ("noleap", "all_leap", "365_day", "366_day", "360_day")
-
-# cftime and datetime classes to use for each calendar name
-datetime_classes = {"default": pydt.datetime, **cftime._cftime.DATE_TYPES}  # noqa
 
 #: Type annotation for strings representing full dates (YYYY-MM-DD), may include time.
 DateStr = NewType("DateStr", str)
@@ -734,492 +723,3 @@ def adapt_clix_meta_yaml(raw: os.PathLike, adapted: os.PathLike):
 
     with open(adapted, "w") as f:
         safe_dump(yml, f)
-
-
-# Datetime-relevant utilities
-
-
-def days_in_year(year: int, calendar: str = "default") -> int:
-    """Return the number of days in the input year according to the input calendar."""
-    return (
-        (datetime_classes[calendar](year + 1, 1, 1) - pydt.timedelta(days=1))
-        .timetuple()
-        .tm_yday
-    )
-
-
-def date_range(
-    *args, calendar: str = "default", **kwargs
-) -> Union[pd.DatetimeIndex, CFTimeIndex]:
-    """Wrap pd.date_range (if calendar == 'default') or xr.cftime_range (otherwise)."""
-    if calendar == "default":
-        return pd.date_range(*args, **kwargs)
-    return xr.cftime_range(*args, calendar=calendar, **kwargs)
-
-
-def select_time(
-    da: Union[xr.DataArray, xr.Dataset],
-    drop: bool = False,
-    season: Union[str, Sequence[str]] = None,
-    month: Union[int, Sequence[int]] = None,
-    doy_bounds: Tuple[int, int] = None,
-    date_bounds: Tuple[str, str] = None,
-):
-    """Select entries according to a time period.
-
-    This conveniently improves xarray's :py:meth:`xarray.DataArray.where` and
-    :py:meth:`xarray.DataArray.sel` with fancier ways of indexing over time elements.
-    In addition to the data `da` and argument `drop`, only one of `season`, `month`,
-    `doy_bounds` or `date_bounds` may be passed.
-
-    Parameters
-    ----------
-    da : xr.DataArray or xr.Dataset
-      Input data.
-    drop: boolean
-      Whether to drop elements outside the period of interest or
-      to simply mask them (default).
-    season: string or sequence of strings
-      One or more of 'DJF', 'MAM', 'JJA' and 'SON'.
-    month: integer or sequence of integers
-      Sequence of month numbers (January = 1 ... December = 12)
-    doy_bounds: 2-tuple of integers
-      The bounds as (start, end) of the period of interest expressed in day-of-year,
-      integers going from 1 (January 1st) to 365 or 366 (December 31st). If calendar
-      awareness is needed, consider using ``date_bounds`` instead.
-      Bounds are inclusive.
-    date_bounds: 2-tuple of strings
-      The bounds as (start, end) of the period of interest expressed as dates in the
-      month-day (%m-%d) format.
-      Bounds are inclusive.
-
-    Returns
-    -------
-    xr.DataArray or xr.Dataset
-      Selected input values. If ``drop=False``, this has the same length as ``da``
-      (along dimension 'time'), but with masked (NaN) values outside the period of
-      interest.
-
-    Examples
-    --------
-    Keep only the values of fall and spring.
-
-    >>> ds = open_dataset("ERA5/daily_surface_cancities_1990-1993.nc")
-    >>> ds.time.size
-    1461
-    >>> out = select_time(ds, drop=True, season=['MAM', 'SON'])
-    >>> out.time.size
-    732
-
-    Or all values between two dates (included).
-
-    >>> out = select_time(ds, drop=True, date_bounds=('02-29', '03-02'))
-    >>> out.time.values
-    array(['1990-03-01T00:00:00.000000000', '1990-03-02T00:00:00.000000000',
-           '1991-03-01T00:00:00.000000000', '1991-03-02T00:00:00.000000000',
-           '1992-02-29T00:00:00.000000000', '1992-03-01T00:00:00.000000000',
-           '1992-03-02T00:00:00.000000000', '1993-03-01T00:00:00.000000000',
-           '1993-03-02T00:00:00.000000000'], dtype='datetime64[ns]')
-    """
-    N = sum(arg is not None for arg in [season, month, doy_bounds, date_bounds])
-    if N > 1:
-        raise ValueError(f"Only one method of indexing may be given, got {N}.")
-
-    if N == 0:
-        return da
-
-    def get_doys(start, end):
-        if start <= end:
-            return np.arange(start, end + 1)
-        return np.concatenate((np.arange(start, 367), np.arange(0, end + 1)))
-
-    if season is not None:
-        if isinstance(season, str):
-            season = [season]
-        mask = da.time.dt.season.isin(season)
-
-    elif month is not None:
-        if isinstance(month, int):
-            month = [month]
-        mask = da.time.dt.month.isin(month)
-
-    elif doy_bounds is not None:
-        mask = da.time.dt.dayofyear.isin(get_doys(*doy_bounds))
-
-    elif date_bounds is not None:
-        # This one is a bit trickier.
-        start, end = date_bounds
-        time = da.time
-        calendar = get_calendar(time)
-        if calendar not in uniform_calendars:
-            # For non-uniform calendars, we can't simply convert dates to doys
-            # conversion to all_leap is safe for all non-uniform calendar as it doesn't remove any date.
-            time = convert_calendar(time, "all_leap")
-            # values of time are the _old_ calendar
-            # and the new calendar is in the coordinate
-            calendar = "all_leap"
-
-        # Get doy of date, this is now safe because the calendar is uniform.
-        doys = get_doys(
-            to_cftime_datetime("2000-" + start, calendar).dayofyr,
-            to_cftime_datetime("2000-" + end, calendar).dayofyr,
-        )
-        mask = time.time.dt.dayofyear.isin(doys)
-        # Needed if we converted calendar, this puts back the correct coord
-        mask["time"] = da.time
-
-    return da.where(mask, drop=drop)
-
-
-def get_calendar(obj: Any, dim: str = "time") -> str:
-    """Return the calendar of an object.
-
-    Parameters
-    ----------
-    obj : Any
-      An object defining some date.
-      If `obj` is an array/dataset with a datetime coordinate, use `dim` to specify its name.
-      Values must have either a datetime64 dtype or a cftime dtype.
-      `obj` can also be a python datetime.datetime, a cftime object or a pandas Timestamp
-      or an iterable of those, in which case the calendar is inferred from the first value.
-    dim : str
-      Name of the coordinate to check (if `obj` is a DataArray or Dataset).
-
-    Raises
-    ------
-    ValueError
-      If no calendar could be inferred.
-
-    Returns
-    -------
-    str
-      The cftime calendar name or "default" when the data is using numpy's or python's datetime types.
-      Will always return "standard" instead of "gregorian", following CF conventions 1.9.
-    """
-    if isinstance(obj, (xr.DataArray, xr.Dataset)):
-        if obj[dim].dtype == "O":
-            obj = obj[dim].where(obj[dim].notnull(), drop=True)[0].item()
-        elif "datetime64" in obj[dim].dtype.name:
-            return "default"
-    elif isinstance(obj, xr.CFTimeIndex):
-        obj = obj.values[0]
-    else:
-        obj = np.take(obj, 0)
-        # Take zeroth element, overcome cases when arrays or lists are passed.
-    if isinstance(obj, pydt.datetime):  # Also covers pandas Timestamp
-        return "default"
-    if isinstance(obj, cftime.datetime):
-        if obj.calendar == "gregorian":
-            return "standard"
-        return obj.calendar
-
-    raise ValueError(f"Calendar could not be inferred from object of type {type(obj)}.")
-
-
-def convert_calendar(
-    source: Union[xr.DataArray, xr.Dataset],
-    target: Union[xr.DataArray, str],
-    align_on: Optional[str] = None,
-    missing: Optional[Any] = None,
-    dim: str = "time",
-) -> Union[xr.DataArray, xr.Dataset]:
-    """Convert a DataArray/Dataset to another calendar using the specified method.
-
-    Only converts the individual timestamps, does not modify any data except in dropping invalid/surplus dates or inserting missing dates.
-
-    If the source and target calendars are either no_leap, all_leap or a standard type, only the type of the time array is modified.
-    When converting to a leap year from a non-leap year, the 29th of February is removed from the array.
-    In the other direction and if `target` is a string, the 29th of February will be missing in the output,
-    unless `missing` is specified, in which case that value is inserted.
-
-    For conversions involving `360_day` calendars, see Notes.
-
-    This method is safe to use with sub-daily data as it doesn't touch the time part of the timestamps.
-
-    Parameters
-    ----------
-    source : xr.DataArray
-      Input array/dataset with a time coordinate of a valid dtype (datetime64 or a cftime.datetime).
-    target : Union[xr.DataArray, str]
-      Either a calendar name or the 1D time coordinate to convert to.
-      If an array is provided, the output will be reindexed using it and in that case, days in `target`
-      that are missing in the converted `source` are filled by `missing` (which defaults to NaN).
-    align_on : {None, 'date', 'year', 'random'}
-      Must be specified when either source or target is a `360_day` calendar, ignored otherwise. See Notes.
-    missing : Optional[any]
-      A value to use for filling in dates in the target that were missing in the source.
-      If `target` is a string, default (None) is not to fill values. If it is an array, default is to fill with NaN.
-    dim : str
-      Name of the time coordinate.
-
-    Returns
-    -------
-    Union[xr.DataArray, xr.Dataset]
-      Copy of source with the time coordinate converted to the target calendar.
-      If `target` is given as an array, the output is reindexed to it, with fill value `missing`.
-      If `target` was a string and `missing` was None (default), invalid dates in the new calendar are dropped, but missing dates are not inserted.
-      If `target` was a string and `missing` was given, then start, end and frequency of the new time axis are inferred and
-      the output is reindexed to that a new array.
-
-    Notes
-    -----
-    If one of the source or target calendars is `360_day`, `align_on` must be specified and two options are offered.
-
-    "year"
-      The dates are translated according to their rank in the year (dayofyear), ignoring their original month and day information,
-      meaning that the missing/surplus days are added/removed at regular intervals.
-
-      From a `360_day` to a standard calendar, the output will be missing the following dates (day of year in parenthesis):
-        To a leap year:
-          January 31st (31), March 31st (91), June 1st (153), July 31st (213), September 31st (275) and November 30th (335).
-        To a non-leap year:
-          February 6th (36), April 19th (109), July 2nd (183), September 12th (255), November 25th (329).
-
-      From standard calendar to a '360_day', the following dates in the source array will be dropped:
-        From a leap year:
-          January 31st (31), April 1st (92), June 1st (153), August 1st (214), September 31st (275), December 1st (336)
-        From a non-leap year:
-          February 6th (37), April 20th (110), July 2nd (183), September 13th (256), November 25th (329)
-
-      This option is best used on daily and subdaily data.
-
-    "date"
-      The month/day information is conserved and invalid dates are dropped from the output. This means that when converting from
-      a `360_day` to a standard calendar, all 31st (Jan, March, May, July, August, October and December) will be missing as there is no equivalent
-      dates in the `360_day` and the 29th (on non-leap years) and 30th of February will be dropped as there are no equivalent dates in
-      a standard calendar.
-
-      This option is best used with data on a frequency coarser than daily.
-
-    "random"
-      Similar to "year", each day of year of the source is mapped to another day of year
-      of the target. However, instead of having always the same missing days according
-      the source and target years, here 5 days are chosen randomly, one for each fifth
-      of the year. However, February 29th is always missing when converting to a leap year,
-      or its value is dropped when converting from a leap year. This is similar to method
-      used in the [LOCA]_ dataset.
-
-      This option best used on daily data.
-
-    References
-    ----------
-    .. [LOCA] Pierce, D. W., D. R. Cayan, and B. L. Thrasher, 2014: Statistical downscaling using Localized Constructed Analogs (LOCA). Journal of Hydrometeorology, volume 15, page 2558-2585
-
-    Examples
-    --------
-    This method does not try to fill the missing dates other than with a constant value,
-    passed with `missing`. In order to fill the missing dates with interpolation, one
-    can simply use xarray's method:
-
-    >>> tas_nl = convert_calendar(tas, 'noleap')  # For the example
-    >>> with_missing = convert_calendar(tas_nl, 'standard', missing=np.NaN)
-    >>> out = with_missing.interpolate_na('time', method='linear')
-
-    Here, if Nans existed in the source data, they will be interpolated too. If that is,
-    for some reason, not wanted, the workaround is to do:
-
-    >>> mask = convert_calendar(tas_nl, 'standard').notnull()
-    >>> out2 = out.where(mask)
-    """
-    cal_src = get_calendar(source, dim=dim)
-
-    if isinstance(target, str):
-        cal_tgt = target
-    else:
-        cal_tgt = get_calendar(target, dim=dim)
-
-    if cal_src == cal_tgt:
-        return source
-
-    if (cal_src == "360_day" or cal_tgt == "360_day") and align_on not in [
-        "year",
-        "date",
-        "random",
-    ]:
-        raise ValueError(
-            "Argument `align_on` must be specified with either 'date', 'year' or "
-            "'random' when converting to or from a '360_day' calendar."
-        )
-    if cal_src != "360_day" and cal_tgt != "360_day":
-        align_on = None
-
-    out = source.copy()
-    # TODO Maybe the 5-6 days to remove could be given by the user?
-    if align_on in ["year", "random"]:
-        if align_on == "year":
-
-            def _yearly_interp_doy(time):
-                # Returns the nearest day in the target calendar of the corresponding "decimal year" in the source calendar
-                yr = int(time.dt.year[0])
-                return np.round(
-                    days_in_year(yr, cal_tgt)
-                    * time.dt.dayofyear
-                    / days_in_year(yr, cal_src)
-                ).astype(int)
-
-            new_doy = source.time.groupby(f"{dim}.year").map(_yearly_interp_doy)
-        elif align_on == "random":
-
-            def _yearly_random_doy(time, rng):
-                # Return a doy in the new calendar, removing the Feb 29th and 5 other
-                # days chosen randomly within 5 sections of 72 days.
-                yr = int(time.dt.year[0])
-                new_doy = np.arange(360) + 1
-                rm_idx = rng.integers(0, 72, 5) + (np.arange(5) * 72)
-                if cal_src == "360_day":
-                    for idx in rm_idx:
-                        new_doy[idx + 1 :] = new_doy[idx + 1 :] + 1
-                    if days_in_year(yr, cal_tgt) == 366:
-                        new_doy[new_doy >= 60] = new_doy[new_doy >= 60] + 1
-                elif cal_tgt == "360_day":
-                    new_doy = np.insert(new_doy, rm_idx - np.arange(5), -1)
-                    if days_in_year(yr, cal_src) == 366:
-                        new_doy = np.insert(new_doy, 60, -1)
-                return new_doy[time.dt.dayofyear - 1]
-
-            new_doy = source.time.groupby(f"{dim}.year").map(
-                _yearly_random_doy, rng=np.random.default_rng()
-            )
-
-        # Convert the source datetimes, but override the doy with our new doys
-        out[dim] = xr.DataArray(
-            [
-                _convert_datetime(datetime, new_doy=doy, calendar=cal_tgt)
-                for datetime, doy in zip(source[dim].indexes[dim], new_doy)
-            ],
-            dims=(dim,),
-            name=dim,
-        )
-        # Remove NaN that where put on invalid dates in target calendar
-        out = out.where(out[dim].notnull(), drop=True)
-        # Remove duplicate timestamps, happens when reducing the number of days
-        out = out.isel({dim: np.unique(out[dim], return_index=True)[1]})
-    else:
-        time_idx = source[dim].indexes[dim]
-        out[dim] = xr.DataArray(
-            [_convert_datetime(time, calendar=cal_tgt) for time in time_idx],
-            dims=(dim,),
-            name=dim,
-        )
-        # Remove NaN that where put on invalid dates in target calendar
-        out = out.where(out[dim].notnull(), drop=True)
-
-    if isinstance(target, str) and missing is not None:
-        target = date_range_like(source[dim], cal_tgt)
-
-    if isinstance(target, xr.DataArray):
-        out = out.reindex({dim: target}, fill_value=missing or np.nan)
-
-    # Copy attrs but change remove `calendar` is still present.
-    out[dim].attrs.update(source[dim].attrs)
-    out[dim].attrs.pop("calendar", None)
-    return out
-
-
-def date_range_like(source: xr.DataArray, calendar: str) -> xr.DataArray:
-    """Generate a datetime array with the same frequency, start and end as another one, but in a different calendar.
-
-    Parameters
-    ----------
-    source : xr.DataArray
-      1D datetime coordinate DataArray
-    calendar : str
-      New calendar name.
-
-    Raises
-    ------
-    ValueError
-      If the source's frequency was not found.
-
-    Returns
-    -------
-    xr.DataArray
-      1D datetime coordinate with the same start, end and frequency as the source, but in the new calendar.
-        The start date is assumed to exist in the target calendar.
-        If the end date doesn't exist, the code tries 1 and 2 calendar days before.
-        Exception when the source is in 360_day and the end of the range is the 30th of a 31-days month,
-        then the 31st is appended to the range.
-    """
-    freq = xr.infer_freq(source)
-    if freq is None:
-        raise ValueError(
-            "`date_range_like` was unable to generate a range as the source frequency was not inferrable."
-        )
-
-    src_cal = get_calendar(source)
-    if src_cal == calendar:
-        return source
-
-    index = source.indexes[source.dims[0]]
-    end_src = index[-1]
-    end = _convert_datetime(end_src, calendar=calendar)
-    if end is np.nan:  # Day is invalid, happens at the end of months.
-        end = _convert_datetime(end_src.replace(day=end_src.day - 1), calendar=calendar)
-        if end is np.nan:  # Still invalid : 360_day to non-leap february.
-            end = _convert_datetime(
-                end_src.replace(day=end_src.day - 2), calendar=calendar
-            )
-    if src_cal == "360_day" and end_src.day == 30 and end.daysinmonth == 31:
-        # For the specific case of daily data from 360_day source, the last day is expected to be "missing"
-        end = end.replace(day=31)
-
-    return xr.DataArray(
-        date_range(
-            _convert_datetime(index[0], calendar=calendar),
-            end,
-            freq=freq,
-            calendar=calendar,
-        ),
-        dims=source.dims,
-        name=source.dims[0],
-    )
-
-
-def _convert_datetime(
-    datetime: Union[pydt.datetime, cftime.datetime],
-    new_doy: Optional[Union[float, int]] = None,
-    calendar: str = "default",
-) -> Union[cftime.datetime, pydt.datetime, float]:
-    """Convert a datetime object to another calendar.
-
-    Nanosecond information are lost as cftime.datetime doesn't support them.
-
-    Parameters
-    ----------
-    datetime: Union[datetime.datetime, cftime.datetime]
-      A datetime object to convert.
-    new_doy:  Optional[Union[float, int]]
-      Allows for redefining the day of year (thus ignoring month and day information from the source datetime).
-      -1 is understood as a nan.
-    calendar: str
-      The target calendar
-
-    Returns
-    -------
-    Union[cftime.datetime, datetime.datetime, np.nan]
-      A datetime object of the target calendar with the same year, month, day and time
-      as the source (month and day according to `new_doy` if given).
-      If the month and day doesn't exist in the target calendar, returns np.nan. (Ex. 02-29 in "noleap")
-    """
-    if new_doy in [np.nan, -1]:
-        return np.nan
-    if new_doy is not None:
-        new_date = cftime.num2date(
-            new_doy - 1,
-            f"days since {datetime.year}-01-01",
-            calendar=calendar if calendar != "default" else "standard",
-        )
-    else:
-        new_date = datetime
-    try:
-        return datetime_classes[calendar](
-            datetime.year,
-            new_date.month,
-            new_date.day,
-            datetime.hour,
-            datetime.minute,
-            datetime.second,
-            datetime.microsecond,
-        )
-    except ValueError:
-        return np.nan
